feat: optimize setExpand and extract event target
